requirements.txt
fastapi
uvicorn
python-dotenv
requests
jinja2
pydantic